Separate interfaces of Features.Wallet project into their own folder.
