+++ Stratis.Bitcoin.Features.LightWallet/LightWalletSyncManager.cs
@@ -2,6 +2,7 @@
 using NBitcoin;
 using Stratis.Bitcoin.Features.Notifications;
 using Stratis.Bitcoin.Features.Wallet;
+using Stratis.Bitcoin.Features.Wallet.Interfaces;
 using Stratis.Bitcoin.Features.Wallet.Notifications;
 using Stratis.Bitcoin.Utilities;
 using System;
