+++ Stratis.Bitcoin.Features.Miner/MiningRPCController.cs
@@ -7,6 +7,7 @@ using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Features.Wallet;
 using Stratis.Bitcoin.Features.RPC.Models;
 using Microsoft.Extensions.Logging;
+using Stratis.Bitcoin.Features.Wallet.Interfaces;
 
 namespace Stratis.Bitcoin.Features.Miner
 {
