+++ Stratis.Bitcoin.Features.Wallet.Tests/WalletManagerTest.cs
@@ -6,6 +6,7 @@ using Newtonsoft.Json;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Connection;
 using Stratis.Bitcoin.Features.MemoryPool;
+using Stratis.Bitcoin.Features.Wallet.Interfaces;
 using Stratis.Bitcoin.Features.Wallet.JsonConverters;
 using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.Tests.Logging;
